requirements.txt
Django==5.2.1
python-dotenv==1.1.0
gunicorn==23.0.0
psycopg[binary]==3.2.9
django-crispy-forms==2.4
django-widget-tweaks==1.5.0
django-markdownx==4.0.9
djangorestframework==3.16.0
crispy-tailwind==1.0.3
requests==2.32.3
Pillow==11.2.1
